+++ packages/cli/src/index.js
@@ -7,6 +7,7 @@ const databaseCli = require('@home-gallery/database/dist/cli');
 const serverCli = require('@home-gallery/server/dist/cli');
 const exportCli = require('@home-gallery/export/dist/cli');
 const interactiveCli = require('./interactive');
+const runCli = require('./run')
 
 const cli = () => {
   yargs.usage('Usage: $0 [global options] <command> [options]')
@@ -16,6 +17,7 @@ const cli = () => {
     .command(serverCli)
     .command(exportCli)
     .command(interactiveCli)
+    .command(runCli)
     .demandCommand()
     .help()
     .alias('h', 'help')